[MODIFY] change data type of real_dist in a_star_node.c to from int to double

+++ inc/a_star_node.h
@@ -6,7 +6,7 @@
 struct node_heap_t
 {
     int vertexName;
-    int real_dist;
+    double real_dist;
     double total_dist_to_src;
     struct node_heap_t *next;
 };
